refactor: multi module 추출으로 인한 삭제
