+++ common-module/src/main/java/org/c4marathon/assignment/model/Account.java
@@ -1,0 +1,13 @@
+package org.c4marathon.assignment.model;
+
+import org.c4marathon.assignment.enums.AccountType;
+
+public interface Account {
+	Long getId();
+
+	Long getBalance();
+
+	AccountType getType();
+
+	String getAccountNumber();
+}
