--- src/main/java/org/c4marathon/assignment/domain/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package org.c4marathon.assignment.domain.repository;
-
-import org.c4marathon.assignment.domain.model.member.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
