--- src/main/java/org/c4marathon/assignment/dto/account/AccountResponseDto.java
@@ -1,9 +1,0 @@
-package org.c4marathon.assignment.dto.account;
-
-import org.c4marathon.assignment.domain.model.account.Account;
-
-public record AccountResponseDto(Long id, Long memberId, Long balance, String accountNumber) {
-	public AccountResponseDto(Account account) {
-		this(account.getId(), account.getMember().getId(), account.getBalance(), account.getAccountNumber());
-	}
-}
