--- src/main/java/org/c4marathon/assignment/dto/member/MoneyDepositRequestDto.java
@@ -1,4 +1,0 @@
-package org.c4marathon.assignment.dto.member;
-
-public record MoneyDepositRequestDto(Long amount) {
-}
